+++ projects/egeo-demo/src/app/demos/barrels.ts
@@ -55,6 +55,7 @@ import { StDraggableCardDemoModule } from './st-draggable-card-demo/st-draggable
 import { StFilterSelectorDemoModule } from './st-filter-selector-demo/st-filter-selector-demo.module';
 import { StZeroPageDemoModule } from './st-zero-page-demo/st-zero-page-demo.module';
 import { StColorPickerDemoModule } from './st-color-picker-demo/st-color-picker-demo.module';
+import { StBubbleOnEllipsisDemoModule } from '@app/demos/st-bubble-on-ellipsis-demo/st-bubble-on-ellipsis-demo.module';
 
 
 export const EGEO_DEMO_MODULES: any[] = [
@@ -62,6 +63,7 @@ export const EGEO_DEMO_MODULES: any[] = [
    StAlertsDemoModule,
    StBreadcrumbsDemoModule,
    StBubbleDemoModule,
+   StBubbleOnEllipsisDemoModule,
    StButtonDemoModule,
    StCheckboxDemoModule,
    StColorPickerDemoModule,